requirements.txt
streamlit
pandas
numpy
python-dotenv
psycopg2-binary
plotly